+++ Changelog.md
@@ -9,7 +9,8 @@
   * New check-site plugin added
   * hacker_news plugin improved. Added two modes.
   * Join to channels with key to irc transport added.
-  * ruby.rb plugin added 
+  * ruby.rb plugin added. Eval simple ruby expression.
+  * ip.py plugin added. Return Ybot external ip address.
 
 ## Ybot-0.2 --> released